+++ NEWS.md
@@ -7,6 +7,9 @@ This package can now also be installed on Windows, where it offers a reduced set
 Most notably, `bettermc::mclapply()` (just like `parallel::mclapply()`) falls back on serial execution, since forking is not available on Windows.
 Still, this allows code employing `bettermc::mclapply()`, which was originally developed for Linux or macOS, to be run on Windows.
 
+### Seeding
+* restore the state of the random number generator (RNG) at the end of `mclapply()` to what it was before; this makes the RNG state of the parent process immune to changes to arguments such as `mc.cores` or `mc.force.fork`
+
 ## Bug Fixes
 * on the first call of `mclapply()` in a session, the environment variable MC_CORES was not respected
 * fix wrong length of affinity.list if `mc.force.fork == TRUE && length(X) == 1`